helios: cast $11 to bigint in landingpage_brand_site_presence upsert
Postgres failed the stock_refresh worker with 'inconsistent types
deduced for parameter $11' because the snapshotId parameter was
used both as a bare value (last_observed_snapshot_id column) and
inside a CASE expression whose other branch is NULL. The NULL
branch left the inferred type ambiguous, so the two positions
disagreed. Cast both uses to bigint to make inference consistent.

Amp-Thread-ID: https://ampcode.com/threads/T-019e329d-ebae-70b6-9d03-73a28144cd11

+++ helios/src/worker/jobs/configWorkersStockRefreshJob.ts
@@ -486,8 +486,8 @@ async function persistSnapshotAndDiff(input: {
             $6, $7, $8,
             $9::timestamptz,
             case when $10 then $9::timestamptz else null end,
-            $11,
-            case when $10 then $11 else null end,
+            $11::bigint,
+            case when $10 then $11::bigint else null end,
             $9::timestamptz
           )
           on conflict (site_dealer_id, brand_id) do update
